requirements.txt
openai==1.54.4
PyMuPDF==1.24.13
python-dotenv==1.0.1
